[Fix] #149 사진 추가 Network Manager 생성
