+++ pophory-iOS/Screen/AddPhoto/ViewController/AddPhotoViewController.swift
@@ -25,6 +25,7 @@ final class AddPhotoViewController: BaseViewController, Navigatable {
     var navigationBarTitleText: String? { return "사진 추가" }
     
     private var presignedURL: PatchPresignedURLRequestDTO?
+    private let networkManager = AddPhotoNetworkManager()
     
     private var albumID: Int?
     private var photoCount: Int?
@@ -63,7 +64,7 @@ final class AddPhotoViewController: BaseViewController, Navigatable {
         
         showNavigationBar()
         setupNavigationBar(with: PophoryNavigationConfigurator.shared)
-        requestGetAlumListAPI()
+        networkManager.requestGetAlumListAPI()
     }
     
     override func viewDidLoad() {
@@ -71,8 +72,8 @@ final class AddPhotoViewController: BaseViewController, Navigatable {
         
         setupTarget()
         setupDelegate()
-        requestGetAlumListAPI()
-        requestGetPresignedURLAPI()
+        networkManager.requestGetAlumListAPI()
+        networkManager.requestGetPresignedURLAPI()
     }
 }
 
@@ -108,16 +109,14 @@ extension AddPhotoViewController {
                           primaryText: "포포리 앨범이 가득찼어요",
                           secondaryText: "아쉽지만,\n다음 업데이트에서 만나요!", firstButtonHandler: goToHome)
             } else {
-//                guard let multipartData = fetchMultiPartData() else { return }
-//                requestPostPhotoAPI(photoInfo: multipartData)
                 if let urlString = presignedURL?.presignedUrl, let url = URL(string: urlString) {
-                        uploadImageToPresignedURL(image: photoImage, presignedURL: url, completion: {_ in
+                    networkManager.uploadImageToPresignedURL(image: photoImage, presignedURL: url, completion: {_ in
                         })
                     } else {
                         print("Invalid URL")
                     }
                 let photoInfo = PostPhotoS3RequestDTO(fileName: presignedURL?.fileName, albumId: albumID, takenAt: dateTaken, studioId: studioID, width: Int(photoImage.size.width), height: Int(photoImage.size.height))
-                requestPostPhotoAPI(photoInfo: photoInfo)
+                networkManager.requestPostPhotoAPI(photoInfo: photoInfo)
                 }
             }
         }
@@ -133,17 +132,6 @@ extension AddPhotoViewController {
     private func setupDelegate() {
         rootView.albumCollectionView.dataSource = self
     }
-    
-//    private func fetchMultiPartData() -> [MultipartFormData]? {
-//        if let imageData = photoImage.jpegData(compressionQuality: 0.8) {
-//            let imageDataProvider = Moya.MultipartFormData(provider: MultipartFormData.FormDataProvider.data(imageData), name: "photo", fileName: "image.jpg", mimeType: "image/jpeg")
-//            guard let albumId = albumID else { return nil }
-//            let albumIDDataProvider = Moya.MultipartFormData(provider: .data("\(albumId)".data(using: .utf8) ?? .empty), name: "albumId")
-//            let dateProvider = Moya.MultipartFormData(provider: .data("\(dateTaken)".data(using: .utf8) ?? .empty), name: "takenAt")
-//            let studioIDProvider = Moya.MultipartFormData(provider: .data("\(studioID)".data(using: .utf8) ?? .empty), name: "studioId")
-//            return [imageDataProvider, albumIDDataProvider, dateProvider, studioIDProvider]
-//        } else { return nil }
-//    }
     
     private func goToHome() {
         dismiss(animated: false)
@@ -195,70 +183,4 @@ extension AddPhotoViewController: DateDataBind, StudioDataBind {
         rootView.studioStackView.setupSelected(selected: true)
         studioID = forIndex
     }
-}
-
-// MARK: - API
-
-extension AddPhotoViewController {
-    func requestGetAlumListAPI() {
-        NetworkService.shared.albumRepository.patchAlbumList() { result in
-            switch result {
-            case .success(let response):
-                self.albumList = response
-            default : return
-            }
-        }
-    }
-    
-    func requestPostPhotoAPI(
-        photoInfo: PostPhotoS3RequestDTO
-    ) {
-        NetworkService.shared.photoRepository.postPhoto(body: photoInfo
-        ) { result in
-            switch result {
-            case .success(_):
-                print("성공")
-                self.goToHome()
-            default : return
-            }
-        }
-    }
-    
-    func requestGetPresignedURLAPI() {
-        NetworkService.shared.photoRepository.patchPresignedPhotoURL( completion: { result in
-            switch result {
-            case .success(let response):
-                self.presignedURL = response
-            default : return
-            }
-        })
-    }
-    
-    func uploadImageToPresignedURL(image: UIImage, presignedURL: URL, completion: @escaping (Error?) -> Void) {
-        guard let imageData = image.jpegData(compressionQuality: 0.8) else {
-            completion(NSError(domain: "com.example.app", code: 0, userInfo: [NSLocalizedDescriptionKey: "Failed to convert image to data"]))
-            return
-        }
-        
-        var request = URLRequest(url: presignedURL)
-        request.httpMethod = "PUT"
-        request.setValue("image/jpeg", forHTTPHeaderField: "Content-Type")
-        request.httpBody = imageData
-        
-        let task = URLSession.shared.dataTask(with: request) { (data, response, error) in
-            if let error = error {
-                completion(error)
-                return
-            }
-            
-            // Check the response status code to ensure successful upload
-            if let httpResponse = response as? HTTPURLResponse,
-                (200...299).contains(httpResponse.statusCode) {
-                completion(nil)  // Upload successful
-                print("업로드 성공")
-            }
-        }
-        task.resume()
-    }
-
 }
